+++ R/internal.R
@@ -250,20 +250,7 @@ no_extra_arguments <- function(...) {
 }
 
 assertthat::on_failure(no_extra_arguments) <- function(call, env) {
-  call_list <- as.list(call)[-1]
-  format_args <- function(i) {
-    if (names(call_list)[i] == "")
-     return(deparse(call_list[[i]]))
-    paste0(names(call_list)[i], "=", deparse(call_list[[i]]))
-  }
-  msg <- paste(vapply(seq_along(call_list), format_args, character(1)),
-               collapse = ", ")
-  if (length(call_list) > 1) {
-    msg <- paste0("unused arguments (", msg, ")")
-  } else {
-    msg <- paste0("unused argument (", msg, ")")
-  }
-  msg
+  "unused arguments"
 }
 
 #' Verify if assertion is met